Add CapturedStreamsMixin for capturing stdout and stderr in tests
